Update MainVerticle.java
Enables the use of external content ( .css, .js ... ) in .html files. Without the fix, external files are not read from webroot ( or a sub-directory)
The fix makes inline CSS in index.html obsolete. CSS embedded in this file can be exteralized and alo be reused in error.html and the like ( code de-duplication)